feat: Widen Tick to u64; from_ticks returns Option<Time>
The master `Tick` counter widens from `u32` to `u64`. With the wider
counter, `time_to_tick`'s `beats: u32 × tick_count: u32` arithmetic
fits in `u64` with ~32 bits of headroom — the `checked_mul().expect()`
backstop is no longer needed. `arb_time` is freed to use
`any::<u32>()` for `beats` (full type domain) instead of the
clamped `0..=100_000` that masked the overflow.

`from_ticks` and `from_ticks_floor` now return `Option<Time>` because
`Time.beats` stays `u32`: `Tick(huge)` values where `huge / chosen_tc`
exceeds `u32::MAX` have no representable `Time`. The `ticks` Conn
unwraps under a documented precondition (`arb_tick` capped at
`u32::MAX × Grid::T1.tick_count()`); runtime callers handle `None`
themselves.

Cast updates ripple through `swing.rs` (i64 → i128 widening),
`sample_tick.rs` (u128 saturation cap → u64::MAX), `transform.rs`,
`scheduler.rs`, `envelope.rs` (linear_u8/smoothstep_u8 widened to
u64), and `cli` test fixtures.

+++ crates/core/src/sync/sample_tick.rs
@@ -88,7 +88,7 @@ impl SampleTickConn {
     }
 
     fn to_tick(x: u128) -> Tick {
-        Tick(x.min(u128::from(u32::MAX)) as u32)
+        Tick(x.min(u128::from(u64::MAX)) as u64)
     }
 }
 
@@ -153,7 +153,7 @@ mod tests {
         #[test]
         fn sample_tick_round_trip(
             stc in arb_integer_stc(),
-            t in 0u32..=1_000_000,
+            t in 0u64..=1_000_000,
         ) {
             let tick = Tick(t);
             let sample = stc.inner(tick);
@@ -193,7 +193,7 @@ mod tests {
         /// the rule.
         #[test]
         fn sample_tick_inner_saturates_on_overflow(
-            tick in (u32::MAX / 2)..=u32::MAX,
+            tick in u64::from(u32::MAX / 2)..=u64::from(u32::MAX),
             bpm_u in 1u32..=100,
             ppqn in 1u32..=8,
         ) {
@@ -360,7 +360,7 @@ mod tests {
             #[test]
             fn stc_samples_per_tick_is_exact_at_48k(
                 bpm in arb_exact_48k_bpm(),
-                t in 0u32..=10_000_000,
+                t in 0u64..=10_000_000,
             ) {
                 let stc = SampleTickConn::new(48_000, Tempo::from_bpm_integer(bpm), PPQN);
                 assert_exact(&stc, Tick(t));
@@ -371,7 +371,7 @@ mod tests {
             #[test]
             fn stc_samples_per_tick_is_exact_at_96k(
                 bpm in arb_exact_96k_bpm(),
-                t in 0u32..=10_000_000,
+                t in 0u64..=10_000_000,
             ) {
                 let stc = SampleTickConn::new(96_000, Tempo::from_bpm_integer(bpm), PPQN);
                 assert_exact(&stc, Tick(t));
@@ -384,7 +384,7 @@ mod tests {
             #[test]
             fn stc_round_trip_identity_48k_96k(
                 (sr, bpm) in arb_exact_sr_bpm(),
-                t in 0u32..=1_000_000,
+                t in 0u64..=1_000_000,
             ) {
                 let stc = SampleTickConn::new(sr, Tempo::from_bpm_integer(bpm), PPQN);
                 let tick = Tick(t);
